# #1758 - faster support function/vector for CartesianProduct/CartesianProductArray