Closes #1758.

The first commit is just formatting (4 spaces instead of 3).

```julia
julia> d = SingleEntryVector(10, 100, 1.0)
julia> X = CartesianProductArray([rand(Ball2, dim=10) for _ in 1:10])
julia> Y = CartesianProduct(rand(Ball2, dim=50), rand(Ball2, dim=50))

julia> @time σ(d, X)
  0.000013 seconds (21 allocations: 3.688 KiB)  # master
  0.000010 seconds (2 allocations: 1.016 KiB)  # this branch

julia> @time ρ(d, X)
  0.000018 seconds (11 allocations: 1.422 KiB)  # master
  0.000010 seconds (1 allocation: 16 bytes)  # this branch

julia> @time σ(d, Y)
  0.000013 seconds (5 allocations: 2.812 KiB)  # master
  0.000013 seconds (2 allocations: 1.359 KiB)  # this branch

julia> @time ρ(d, Y)
  0.000011 seconds (3 allocations: 1008 bytes)  # master
  0.000005 seconds (1 allocation: 16 bytes)  # this branch
```